refactor: remove remaining extension core imports
