+++ extensions/imessage/src/shared.ts
@@ -3,19 +3,17 @@ import {
   collectAllowlistProviderRestrictSendersWarnings,
 } from "openclaw/plugin-sdk/channel-policy";
 import {
+  buildChannelConfigSchema,
+  DEFAULT_ACCOUNT_ID,
   deleteAccountFromConfigSection,
-  setAccountEnabledInConfigSection,
-} from "../../../src/channels/plugins/config-helpers.js";
-import { buildChannelConfigSchema } from "../../../src/channels/plugins/config-schema.js";
-import type { ChannelPlugin } from "../../../src/channels/plugins/types.plugin.js";
-import { getChatChannelMeta } from "../../../src/channels/registry.js";
-import { IMessageConfigSchema } from "../../../src/config/zod-schema.providers-core.js";
-import {
   formatTrimmedAllowFromEntries,
+  getChatChannelMeta,
+  IMessageConfigSchema,
   resolveIMessageConfigAllowFrom,
   resolveIMessageConfigDefaultTo,
-} from "../../../src/plugin-sdk/channel-config-helpers.js";
-import { DEFAULT_ACCOUNT_ID } from "../../../src/routing/session-key.js";
+  setAccountEnabledInConfigSection,
+  type ChannelPlugin,
+} from "openclaw/plugin-sdk/imessage";
 import {
   listIMessageAccountIds,
   resolveDefaultIMessageAccountId,
