+++ extensions/signal/src/setup-surface.ts
@@ -1,6 +1,9 @@
-import { setSetupChannelEnabled, type ChannelSetupWizard } from "openclaw/plugin-sdk/setup";
-import { detectBinary } from "../../../src/plugins/setup-binary.js";
-import { installSignalCli } from "../../../src/plugins/signal-cli-install.js";
+import {
+  detectBinary,
+  installSignalCli,
+  setSetupChannelEnabled,
+  type ChannelSetupWizard,
+} from "openclaw/plugin-sdk/setup";
 import { listSignalAccountIds, resolveSignalAccount } from "./accounts.js";
 import {
   createSignalCliPathTextInput,
