+++ extensions/slack/src/setup-surface.ts
@@ -1,4 +1,5 @@
 import {
+  formatDocsLink,
   noteChannelLookupFailure,
   noteChannelLookupSummary,
   type OpenClawConfig,
@@ -11,7 +12,6 @@ import type {
   ChannelSetupWizard,
   ChannelSetupWizardAllowFromEntry,
 } from "openclaw/plugin-sdk/setup";
-import { formatDocsLink } from "../../../src/terminal/links.js";
 import { resolveDefaultSlackAccountId, resolveSlackAccount } from "./accounts.js";
 import { resolveSlackChannelAllowlist } from "./resolve-channels.js";
 import { resolveSlackUserAllowlist } from "./resolve-users.js";
