+++ extensions/telegram/src/setup-core.ts
@@ -1,6 +1,8 @@
 import {
   createEnvPatchedAccountSetupAdapter,
   DEFAULT_ACCOUNT_ID,
+  formatCliCommand,
+  formatDocsLink,
   patchChannelConfigForAccount,
   promptResolvedAllowFrom,
   splitSetupEntries,
@@ -8,8 +10,6 @@ import {
   type WizardPrompter,
 } from "openclaw/plugin-sdk/setup";
 import type { ChannelSetupAdapter, ChannelSetupDmPolicy } from "openclaw/plugin-sdk/setup";
-import { formatCliCommand } from "../../../src/cli/command-format.js";
-import { formatDocsLink } from "../../../src/terminal/links.js";
 import { resolveDefaultTelegramAccountId, resolveTelegramAccount } from "./accounts.js";
 import { fetchTelegramChatId } from "./api-fetch.js";
