+++ extensions/telegram/src/shared.ts
@@ -3,12 +3,14 @@ import {
   createScopedAccountConfigAccessors,
   createScopedChannelConfigBase,
 } from "openclaw/plugin-sdk/channel-config-helpers";
-import { buildChannelConfigSchema } from "../../../src/channels/plugins/config-schema.js";
-import type { ChannelPlugin } from "../../../src/channels/plugins/types.plugin.js";
-import { getChatChannelMeta } from "../../../src/channels/registry.js";
-import type { OpenClawConfig } from "../../../src/config/config.js";
-import { TelegramConfigSchema } from "../../../src/config/zod-schema.providers-core.js";
-import { normalizeAccountId } from "../../../src/routing/session-key.js";
+import {
+  buildChannelConfigSchema,
+  getChatChannelMeta,
+  normalizeAccountId,
+  TelegramConfigSchema,
+  type ChannelPlugin,
+  type OpenClawConfig,
+} from "openclaw/plugin-sdk/telegram";
 import { inspectTelegramAccount } from "./account-inspect.js";
 import {
   listTelegramAccountIds,
