+++ extensions/whatsapp/src/setup-surface.ts
@@ -1,6 +1,8 @@
 import path from "node:path";
 import {
   DEFAULT_ACCOUNT_ID,
+  formatCliCommand,
+  formatDocsLink,
   normalizeAccountId,
   normalizeAllowFromEntries,
   normalizeE164,
@@ -11,8 +13,6 @@ import {
   type OpenClawConfig,
 } from "openclaw/plugin-sdk/setup";
 import type { ChannelSetupWizard } from "openclaw/plugin-sdk/setup";
-import { formatCliCommand } from "../../../src/cli/command-format.js";
-import { formatDocsLink } from "../../../src/terminal/links.js";
 import { listWhatsAppAccountIds, resolveWhatsAppAuthDir } from "./accounts.js";
 import { loginWeb } from "./login.js";
 import { whatsappSetupAdapter } from "./setup-core.js";
