+++ src/plugin-sdk/slack.ts
@@ -81,3 +81,5 @@ export {
 } from "../../extensions/slack/src/actions.js";
 export { recordSlackThreadParticipation } from "../../extensions/slack/src/sent-thread-cache.js";
 export { handleSlackMessageAction } from "./slack-message-actions.js";
+export { createSlackActions } from "../channels/plugins/slack.actions.js";
+export type { SlackActionContext } from "../agents/tools/slack-actions.js";
